% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCF-package.R
\docType{package}
\name{spCF}
\alias{spCF}
\alias{spCF-package}
\title{spCF: Coarse-to-Fine Spatial Modeling}
\description{
Provides functions for coarse-to-fine spatial modeling (CFSM),
enabling fast spatial prediction, regression, and uncertainty
quantification. Suitable for moderate to large samples.
}
\author{
\strong{Maintainer}: Daisuke Murakami \email{dmuraka@ism.ac.jp}

Authors:
\itemize{
  \item Alexis Comber
  \item Takahiro Yoshida
  \item Narumasa Tsutsumida
  \item Chris Brunsdon
  \item Tomoki Nakaya
}

}
\keyword{internal}
